Break cyclic import
